node: rename components to pipeline
