+++ node/BUILD
@@ -52,7 +52,7 @@ cc_library(
         ":neighbor_shared",
         ":node_shared",
         "//common/network:uri",
-        "//node/components:transaction_requester",
+        "//node/pipeline:transaction_requester",
         "//node/services:tcp_sender",
         "//node/services:udp_sender",
         "//utils/handles:rand",
@@ -112,14 +112,14 @@ cc_library(
         ":neighbor_shared",
         ":recent_seen_bytes_cache",
         ":tips_cache",
-        "//node/components:broadcaster_shared",
-        "//node/components:processor_shared",
-        "//node/components:receiver_shared",
-        "//node/components:responder_shared",
-        "//node/components:tips_requester_shared",
-        "//node/components:tips_solidifier_shared",
-        "//node/components:transaction_requester_shared",
-        "//node/components:transaction_requester_worker_shared",
+        "//node/pipeline:broadcaster_shared",
+        "//node/pipeline:processor_shared",
+        "//node/pipeline:receiver_shared",
+        "//node/pipeline:responder_shared",
+        "//node/pipeline:tips_requester_shared",
+        "//node/pipeline:tips_solidifier_shared",
+        "//node/pipeline:transaction_requester_shared",
+        "//node/pipeline:transaction_requester_worker_shared",
     ],
 )
 
@@ -129,13 +129,13 @@ cc_library(
     deps = [
         ":node_shared",
         "//ciri:core_shared",
-        "//node/components:broadcaster",
-        "//node/components:processor",
-        "//node/components:receiver",
-        "//node/components:responder",
-        "//node/components:tips_requester",
-        "//node/components:tips_solidifier",
-        "//node/components:transaction_requester",
-        "//node/components:transaction_requester_worker",
+        "//node/pipeline:broadcaster",
+        "//node/pipeline:processor",
+        "//node/pipeline:receiver",
+        "//node/pipeline:responder",
+        "//node/pipeline:tips_requester",
+        "//node/pipeline:tips_solidifier",
+        "//node/pipeline:transaction_requester",
+        "//node/pipeline:transaction_requester_worker",
     ],
 )
