+++ node/node.h
@@ -9,15 +9,15 @@
 #define __NODE_NODE_H__
 
 #include "common/errors.h"
-#include "node/components/broadcaster.h"
-#include "node/components/processor.h"
-#include "node/components/receiver.h"
-#include "node/components/responder.h"
-#include "node/components/tips_requester.h"
-#include "node/components/tips_solidifier.h"
-#include "node/components/transaction_requester.h"
-#include "node/components/transaction_requester_worker.h"
 #include "node/neighbor.h"
+#include "node/pipeline/broadcaster.h"
+#include "node/pipeline/processor.h"
+#include "node/pipeline/receiver.h"
+#include "node/pipeline/responder.h"
+#include "node/pipeline/tips_requester.h"
+#include "node/pipeline/tips_solidifier.h"
+#include "node/pipeline/transaction_requester.h"
+#include "node/pipeline/transaction_requester_worker.h"
 #include "node/recent_seen_bytes_cache.h"
 #include "node/tips_cache.h"
 #include "utils/handles/rw_lock.h"
